색칠 도중 결과가 스테이지 선택 화면에서 보이도록 하기

+++ Assets/Scripts/MainGame.cs
@@ -84,8 +84,13 @@ public class MainGame : MonoBehaviour {
         pinchZoom.ResetZoom();
     }
 
+    int resetCount;
+
     public void ResetStage() {
-        gridWorld.DeleteSaveFileAndReloadScene();
+        resetCount++;
+        if (resetCount > 5) {
+            gridWorld.DeleteSaveFileAndReloadScene();
+        }
     }
 
     public void LoadStageSelectionScene() {
